+++ code/src/apps/sdl_map_example/map_view.cc
@@ -5,21 +5,25 @@ MapView::MapView(SDL_Window *window, int width, int height, int zoom): window(wi
 	resize(width, height);
 	update_bounds();
 }
+
 void MapView::center_coords(double lat, double lng){
 	double x = (lng + 180.0f) / 360.0f;
 	double y = (1.0f - log(tan(lat * M_PI / 180.0f) + 1.0f / cos(lat * M_PI / 180.0f)) / M_PI) / 2.0f;
-	offsetx = x * (TILESIZE << zoom) - width / 2;
-	offsety = y * (TILESIZE << zoom) - height / 2;
+	offsetx = static_cast<int32_t>(x * (TILESIZE << zoom) - width / 2);
+	offsety = static_cast<int32_t>(y * (TILESIZE << zoom) - height / 2);
 }
+
 void MapView::render(){
 	SDL_Surface *screen = SDL_GetWindowSurface(window);
 	tiles.render(screen, offsetx, offsety);
 	SDL_UpdateWindowSurface(window);
 }
+
 void MapView::move_by(int dx, int dy){
 	offsetx += dx;
 	offsety += dy;
 }
+
 void MapView::zoom_at(int x, int y){
 	if(zoom < 22){
 		zoom++;
@@ -27,9 +31,11 @@ void MapView::zoom_at(int x, int y){
 		offsety = offsety * 2 + y;
 	}
 }
+
 void MapView::zoom_in(){
 	zoom_at(width / 2, height / 2);
 }
+
 void MapView::zoom_out(){
 	if(zoom > 3){
 		zoom--;
@@ -37,6 +43,7 @@ void MapView::zoom_out(){
 		offsety = (offsety - height / 2) / 2;
 	}
 }
+
 void MapView::resize(int w, int h){
 	/* keep the same center*/
 	offsetx += (width - w) / 2;
@@ -45,6 +52,7 @@ void MapView::resize(int w, int h){
 	width = w;
 	height = h;
 }
+
 void MapView::update_bounds(){
 	int maxy = (1 << zoom) * TILESIZE - height;
 	if(offsety < 0)
